moved pages into different folders, and changed some letters to uppercase for routing errors
