+++ src/Components/pages/admin/Inventories.js
