--- src/pages/Shop/Products.js
@@ -1,3 +1,0 @@
-// Compoenents
-// Item
-    // Checkout comp.
